LibJS: Enable opting into GDB-supported JIT compilation
Now the JIT compiler can generate a NativeExecutable with an ELF format
supported by the GDB JIT Interface, instead of a raw, mmap'd code page.

This is opt-in through a `LIBJS_JIT_GDB` environment variable. This is
because it may increase the memory footprint significantly, specially as
we fill the (currently barebones) ELF image with relevant symbols &
debug information.